scripts/requirements.txt
pandas>=1.3.0
elasticsearch>=7.0.0,<8.0.0
python-dotenv>=0.19.0
